% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Emulate_Trial_Bootstrap.R
\name{Emulate_Trial_Bootstrap}
\alias{Emulate_Trial_Bootstrap}
\title{Bootstrap confidence intervals of the estimates obtained from Emulate_Trial.}
\usage{
Emulate_Trial_Bootstrap(
  data,
  grace,
  time,
  formula_RHS = NULL,
  var_list,
  use.weights = T,
  WC_limit = NULL,
  ncores = 1,
  mboot,
  CI_type = "gaussian",
  a = 95
)
}
\arguments{
\item{data}{A wide format data frame}

\item{grace}{The length of the grace period.}

\item{time}{Fixed time point to assess survival and survival difference.}

\item{formula_RHS}{Right-hand side of formula for the relationship between
if and when a patient is treated and his covariates.
Is used to model the artifical censoring to obtain weights.}

\item{var_list}{A named list naming the variables in data that should be used
time_OS, status_OS, time_treatment, status_treatment.}

\item{use.weights}{True false indicating whether the trial emulation should use
weights to adjust for the selection bias/informative censoring introduced by the
artificial censoring from the cloning strategy.}

\item{WC_limit}{The Webster-Clark limit to specify the which patients should be weigthed in
the experimental arm based on time_treatment. Only those treated after the WC_limit are
weighed. If NULL, all patients are weighted.}

\item{ncores}{Number of cores used for parrelized computations.}

\item{mboot}{Number of bootstrap repitions used.}

\item{CI_type}{The choice of confidence interval used, either "none", "normal" or "quantile"}

\item{a}{corresponding the an a\%-confidence interval.}
}
\value{
Confidence intervals of the survival estimates
}
\description{
Bootstrap confidence intervals of the estimates obtained from Emulate_Trial.
}
\examples{
library(survival)
library(dplyr)
library(CCWTrialEmulation)


Emulate_Trial(
 data = jasa |> mutate(
   futime = ifelse(futime == 0, 1, futime),
   #does not accept events at time 0
   wait.time = ifelse(wait.time == 0, 1, wait.time)
 ),
 grace = 60,
 time = 1 * 365,
 formula_RHS = "~ age + surgery",
 var_list = list(
   "time_OS" = "futime",
   "status_OS" = "fustat",
   "time_treatment" = "wait.time",
   "status_treatment" = "transplant"
 )
)

Emulate_Trial_Bootstrap(
 jasa |> mutate(
   futime = ifelse(futime == 0, 1, futime),
   #does not accept events at time 0
   wait.time = ifelse(wait.time == 0, 1, wait.time)
 ),
 grace = 60,
 time = 1*365,
 formula_RHS = "~ age + surgery",
 var_list = list(
   "time_OS" = "futime",
   "status_OS" = "fustat",
   "time_treatment" = "wait.time",
   "status_treatment" = "transplant"
 ),
 use.weights = T,
 ncores = 5,
 mboot = 100,
 CI_type = "gaussian",
 a = 95
)




}
